backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
motor==3.3.2
pymongo==4.6.2
pydantic==2.9.1
pydantic-settings==2.6.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pytesseract==0.3.13
Pillow==10.4.0
bcrypt==4.1.3
pytest==8.3.3
pytest-asyncio==0.23.7
httpx[http2]==0.27.2
python-dotenv==1.0.1
